refactor: view 관련 코드 views 디렉토리 안으로 이동
